🚀 Migrate Next.js from App Router to Pages Router & Fix Blog Integration
- ✅ Removed `src/app/` folder (migrated to pages router)
- ✅ Added `_app.js` in `src/pages/` for consistent layout handling
- ✅ Moved `layout.js` and `page.js` functionality into `_app.js` and `index.js`
- ✅ Updated `next.config.js`:
  - Wrapped `rewrites()` to apply only in dev mode
  - Ensured `next export` works correctly
- ✅ Fixed `package.json` scripts to support Astro static build in `out/blog/`
